+++ Assets/Scripts/MaskBarScript.cs
@@ -7,13 +7,13 @@ public class MaskBarScript : MonoBehaviour
 {
     public Slider slider;
 
-    public void SetMaxMask(int maskValue)
+    public void SetMaxMask(float maskValue)
     {
         slider.maxValue = maskValue;
         slider.value = maskValue;
     }
 
-    public void SetMask(int maskValue)
+    public void SetMask(float maskValue)
     {
         slider.value = maskValue;
     }
